requirements.txt
chardet==4.0.0
paramiko==2.9.2
pip-chill==1.0.1
PyDrive==1.3.1
psycopg2-binary==2.9.3
python-dotenv==0.19.2
selenium==3.141.0
